+++ Frontend/src/components/body/Body.tsx
